pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.51.0"
python-dotenv = "^1.0.1"
discord = "^2.3.2"
pynacl = "^1.5.0"
pandas = "^2.2.3"
langchain = "^0.3.1"
langchain-core = "^0.3.7"
langchain-community = "^0.3.1"
discord-py = {extras = ["voice"], version = "^2.4.0"}
langchain-openai = "^0.2.1"
deepl = "^1.19.1"
pillow = "^10.4.0"
dependency-injector = "^4.42.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
